Fix getMulti causing PHP warning.
It was possible is one of the alloptions was not previously found in the cache it could throw a warning. this is ultimately because getMulti would return a internal_key => value map, so there was no easy way to dedupe for array_combine. Instead we have getMulti return an array of key => value map which is much better.