Make "broadly not recommended" cache flush warning filterable
In local development environments it is frequently desirable to flush
the cache manually after making certain database or content updates,
and many WP-CLI commands (including installation and content imports)
call wp_cache_flush() automatically after completing their updates.

To avoid verbosely and repeatedly telling developers that the thing they
wanted to do in their local is wrong, this PR introduces a filter
"pecl_memcached/warn_on_flush" to permit a local environment to hook
"__return_false" in to suppress this warning when it is not needed.